# [target allocator] Use a uint64 hash instead of a string