--- a/cmd/otel-allocator/internal/diff/diff.go
+++ b/cmd/otel-allocator/internal/diff/diff.go
@@ -5,33 +5,33 @@ package diff
 
 // Changes is the result of the difference between two maps – items that are added and items that are removed
 // This map is used to reconcile state differences.
-type Changes[T Hasher] struct {
-	additions map[string]T
-	removals  map[string]T
+type Changes[K comparable, T Hasher[K]] struct {
+	additions map[K]T
+	removals  map[K]T
 }
 
-type Hasher interface {
-	Hash() string
+type Hasher[K comparable] interface {
+	Hash() K
 }
 
-func NewChanges[T Hasher](additions map[string]T, removals map[string]T) Changes[T] {
-	return Changes[T]{additions: additions, removals: removals}
+func NewChanges[K comparable, T Hasher[K]](additions map[K]T, removals map[K]T) Changes[K, T] {
+	return Changes[K, T]{additions: additions, removals: removals}
 }
 
-func (c Changes[T]) Additions() map[string]T {
+func (c Changes[K, T]) Additions() map[K]T {
 	return c.additions
 }
 
-func (c Changes[T]) Removals() map[string]T {
+func (c Changes[K, T]) Removals() map[K]T {
 	return c.removals
 }
 
 // Maps generates Changes for two maps with the same type signature by checking for any removals and then checking for
 // additions.
 // TODO: This doesn't need to create maps, it can return slices only. This function doesn't need to insert the values.
-func Maps[T Hasher](current, new map[string]T) Changes[T] {
-	additions := map[string]T{}
-	removals := map[string]T{}
+func Maps[K comparable, T Hasher[K]](current, new map[K]T) Changes[K, T] {
+	additions := map[K]T{}
+	removals := map[K]T{}
 	for key, newValue := range new {
 		if currentValue, found := current[key]; !found {
 			additions[key] = newValue
@@ -45,7 +45,7 @@ func Maps[T Hasher](current, new map[string]T) Changes[T] {
 			removals[key] = value
 		}
 	}
-	return Changes[T]{
+	return Changes[K, T]{
 		additions: additions,
 		removals:  removals,
 	}