We use a (relatively long) string as our target hash for no good reason in the target allocator. Since what we're really hashing are labels, and these come with a built-in hash method that returns a uint64, we use that instead. This gives some impressive performance benefits for such a simple change:

```
goos: linux
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator
cpu: AMD Ryzen 9 7950X3D 16-Core Processor          
                                                             │ bench_main.txt │        bench_branch_hash.txt        │
                                                             │     sec/op     │   sec/op     vs base                │
ProcessTargets/per-node/1000-32                                   1.058m ± 3%   1.015m ± 3%   -4.12% (p=0.000 n=10)
ProcessTargets/least-weighted/1000-32                             1.064m ± 0%   1.015m ± 1%   -4.60% (p=0.000 n=10)
ProcessTargets/consistent-hashing/1000-32                         1.068m ± 2%   1.065m ± 4%        ~ (p=0.393 n=10)
ProcessTargets/least-weighted/10000-32                            7.564m ± 1%   7.186m ± 1%   -5.00% (p=0.000 n=10)
ProcessTargets/consistent-hashing/10000-32                        7.594m ± 1%   7.007m ± 2%   -7.73% (p=0.000 n=10)
ProcessTargets/per-node/10000-32                                  7.582m ± 3%   6.916m ± 3%   -8.79% (p=0.000 n=10)
ProcessTargets/least-weighted/100000-32                           82.35m ± 3%   70.92m ± 4%  -13.88% (p=0.000 n=10)
ProcessTargets/consistent-hashing/100000-32                       81.84m ± 4%   68.53m ± 2%  -16.26% (p=0.000 n=10)
ProcessTargets/per-node/100000-32                                 80.45m ± 3%   70.13m ± 3%  -12.83% (p=0.000 n=10)
ProcessTargets/least-weighted/800000-32                           875.1m ± 5%   709.9m ± 7%  -18.88% (p=0.000 n=10)
ProcessTargets/consistent-hashing/800000-32                       882.4m ± 6%   703.9m ± 9%  -20.22% (p=0.000 n=10)
ProcessTargets/per-node/800000-32                                 889.6m ± 3%   708.0m ± 5%  -20.41% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/1000-32            1.497m ± 1%   1.499m ± 1%        ~ (p=0.796 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/1000-32        1.473m ± 4%   1.498m ± 1%        ~ (p=0.684 n=10)
ProcessTargetsWithRelabelConfig/per-node/1000-32                  1.526m ± 3%   1.522m ± 1%        ~ (p=0.853 n=10)
ProcessTargetsWithRelabelConfig/per-node/10000-32                 11.72m ± 2%   11.00m ± 2%   -6.10% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/10000-32           11.88m ± 1%   10.90m ± 1%   -8.26% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/10000-32       11.85m ± 1%   10.83m ± 1%   -8.57% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/100000-32          116.6m ± 2%   110.1m ± 4%   -5.61% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/100000-32      116.4m ± 2%   110.1m ± 3%   -5.41% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/100000-32                114.8m ± 2%   108.9m ± 3%   -5.12% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/800000-32          972.9m ± 4%   870.3m ± 3%  -10.54% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/800000-32      957.1m ± 3%   874.8m ± 2%   -8.60% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/800000-32               1007.6m ± 2%   870.1m ± 3%  -13.65% (p=0.000 n=10)
geomean                                                           32.23m        29.43m        -8.68%

                                                             │ bench_main.txt │        bench_branch_hash.txt        │
                                                             │      B/op      │     B/op      vs base               │
ProcessTargets/per-node/1000-32                                  4.249Mi ± 0%   4.164Mi ± 0%  -2.00% (p=0.000 n=10)
ProcessTargets/least-weighted/1000-32                            4.249Mi ± 0%   4.164Mi ± 0%  -2.00% (p=0.000 n=10)
ProcessTargets/consistent-hashing/1000-32                        4.249Mi ± 0%   4.164Mi ± 0%  -2.00% (p=0.000 n=10)
ProcessTargets/least-weighted/10000-32                           42.39Mi ± 0%   41.57Mi ± 0%  -1.94% (p=0.000 n=10)
ProcessTargets/consistent-hashing/10000-32                       42.39Mi ± 0%   41.57Mi ± 0%  -1.94% (p=0.000 n=10)
ProcessTargets/per-node/10000-32                                 42.39Mi ± 0%   41.57Mi ± 0%  -1.95% (p=0.000 n=10)
ProcessTargets/least-weighted/100000-32                          424.0Mi ± 0%   415.6Mi ± 0%  -1.98% (p=0.000 n=10)
ProcessTargets/consistent-hashing/100000-32                      423.9Mi ± 0%   415.6Mi ± 0%  -1.97% (p=0.000 n=10)
ProcessTargets/per-node/100000-32                                423.9Mi ± 0%   415.6Mi ± 0%  -1.97% (p=0.000 n=10)
ProcessTargets/least-weighted/800000-32                          3.356Gi ± 0%   3.277Gi ± 0%  -2.35% (p=0.000 n=10)
ProcessTargets/consistent-hashing/800000-32                      3.356Gi ± 0%   3.277Gi ± 0%  -2.35% (p=0.000 n=10)
ProcessTargets/per-node/800000-32                                3.356Gi ± 0%   3.277Gi ± 0%  -2.35% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/1000-32           4.502Mi ± 0%   4.452Mi ± 0%  -1.12% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/1000-32       4.502Mi ± 0%   4.452Mi ± 0%  -1.12% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/1000-32                 4.502Mi ± 0%   4.452Mi ± 0%  -1.12% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/10000-32                44.97Mi ± 0%   44.48Mi ± 0%  -1.09% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/10000-32          44.97Mi ± 0%   44.48Mi ± 0%  -1.09% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/10000-32      44.97Mi ± 0%   44.48Mi ± 0%  -1.09% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/100000-32         449.9Mi ± 0%   444.8Mi ± 0%  -1.12% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/100000-32     449.9Mi ± 0%   444.8Mi ± 0%  -1.12% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/100000-32               449.9Mi ± 0%   444.8Mi ± 0%  -1.12% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/800000-32         3.535Gi ± 1%   3.489Gi ± 0%  -1.28% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/800000-32     3.535Gi ± 0%   3.489Gi ± 0%  -1.28% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/800000-32               3.561Gi ± 1%   3.489Gi ± 0%  -2.01% (p=0.000 n=10)
geomean                                                          131.0Mi        128.8Mi       -1.64%

                                                             │ bench_main.txt │        bench_branch_hash.txt        │
                                                             │   allocs/op    │  allocs/op   vs base                │
ProcessTargets/per-node/1000-32                                   5.311k ± 0%   3.312k ± 0%  -37.65% (p=0.000 n=10)
ProcessTargets/least-weighted/1000-32                             5.311k ± 0%   3.311k ± 0%  -37.66% (p=0.000 n=10)
ProcessTargets/consistent-hashing/1000-32                         5.311k ± 0%   3.311k ± 0%  -37.66% (p=0.000 n=10)
ProcessTargets/least-weighted/10000-32                            52.97k ± 0%   32.97k ± 0%  -37.76% (p=0.000 n=10)
ProcessTargets/consistent-hashing/10000-32                        52.97k ± 0%   32.97k ± 0%  -37.76% (p=0.000 n=10)
ProcessTargets/per-node/10000-32                                  52.97k ± 0%   32.97k ± 0%  -37.76% (p=0.000 n=10)
ProcessTargets/least-weighted/100000-32                           528.5k ± 0%   328.5k ± 0%  -37.85% (p=0.000 n=10)
ProcessTargets/consistent-hashing/100000-32                       528.5k ± 0%   328.5k ± 0%  -37.85% (p=0.000 n=10)
ProcessTargets/per-node/100000-32                                 528.5k ± 0%   328.5k ± 0%  -37.85% (p=0.000 n=10)
ProcessTargets/least-weighted/800000-32                           4.230M ± 0%   2.630M ± 0%  -37.82% (p=0.000 n=10)
ProcessTargets/consistent-hashing/800000-32                       4.230M ± 0%   2.630M ± 0%  -37.82% (p=0.000 n=10)
ProcessTargets/per-node/800000-32                                 4.230M ± 0%   2.630M ± 0%  -37.82% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/1000-32            7.303k ± 0%   6.303k ± 0%  -13.69% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/1000-32        7.303k ± 0%   6.303k ± 0%  -13.69% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/1000-32                  7.303k ± 0%   6.303k ± 0%  -13.69% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/10000-32                 72.93k ± 0%   62.93k ± 0%  -13.71% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/10000-32           72.93k ± 0%   62.93k ± 0%  -13.71% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/10000-32       72.93k ± 0%   62.93k ± 0%  -13.71% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/100000-32          728.3k ± 0%   628.3k ± 0%  -13.73% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/100000-32      728.3k ± 0%   628.3k ± 0%  -13.73% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/100000-32                728.3k ± 0%   628.3k ± 0%  -13.73% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/800000-32          5.827M ± 0%   5.027M ± 0%  -13.73% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/800000-32      5.827M ± 0%   5.027M ± 0%  -13.73% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/800000-32                5.829M ± 0%   5.027M ± 0%  -13.76% (p=0.000 n=10)
geomean                                                           185.8k        136.2k       -26.72%

```